fix(content-releases): fix reviews
